+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon server.js & webpack-dev-server",
-    "dev": "nodemon server.js",
+    "dev": "nodemon server.ts",
     "build": "webpack --mode production"
   },
   "author": "Mobius Collective",
@@ -19,11 +19,7 @@
     "d3": "^7.9.0",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
-    "jsonwebtoken": "^9.0.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "ts-node": "^10.9.2",
-    "path": "^0.12.7"
+    "nodemon": "^3.1.4"
   },
   "devDependencies": {
     "@babel/core": "^7.24.7",
